requirements.txt
telethon>=1.35.0
pyyaml>=6.0
requests>=2.31.0